refactor(api): split engine/api.rs into focused submodules (#565) (#610)
Break the 1606-line src/engine/api.rs into a directory module under the
800-line ceiling enforced by CLAUDE.md and the project's coding-style rule:

- src/engine/api/mod.rs           module wiring + public re-exports
- src/engine/api/types.rs         NAPI parameter/result structs
- src/engine/api/engine.rs        ImageEngine def + constructors + lazy
                                  ICC/EXIF accessors
- src/engine/api/pipeline_ops.rs  resize/crop/rotate/flip/grayscale/...
                                  /preset operation chain
- src/engine/api/output_ops.rs    toBuffer/toFile and metric variants
- src/engine/api/batch_ops.rs     dimensions/hasIccProfile/processBatch
- src/engine/api/tests.rs         relocated inline unit test

No behaviour change: napi codegen output (index.d.ts) regenerates
identically aside from re-emission ordering. All public paths
(crate::engine::api::ImageEngine etc.) remain stable thanks to the
re-exports in mod.rs.

Verification:
- cargo +1.88.0 fmt --check
- cargo +1.88.0 clippy --no-default-features -- -D warnings
- cargo test --no-default-features --no-fail-fast (443 tests pass)
- npm run build && npm test (green)
