--- src/engine/api.rs
@@ -1,1606 +1,0 @@
-#![cfg_attr(not(feature = "napi"), allow(dead_code))]
-
-// src/engine/api.rs
-//
-// ImageEngine structure and NAPI implementation.
-// This is the main public API for the image processing engine.
-
-// BatchResult is used in ts_return_type attribute (line 446) - compiler can't detect this
-// Source is used via Source::Memory and Source::Mapped
-use super::firewall::FirewallConfig;
-#[cfg(feature = "napi")]
-use super::firewall::FirewallPolicy;
-#[allow(unused_imports)]
-use crate::engine::io::{extract_exif_raw, extract_icc_profile_lossy, load_file_safe, Source};
-#[cfg(feature = "napi")]
-#[allow(unused_imports)]
-use crate::engine::tasks::{
-    BatchResult, BatchTask, BatchWithMetricsTask, EncodeTargetBytesTask, EncodeTask,
-    EncodeWithMetricsTask, TaskContext, UnifiedEncodeTask, UnifiedEncodeToFileTask, WriteFileTask,
-    WriteFileWithMetricsTask,
-};
-#[cfg(feature = "napi")]
-use crate::error::LazyImageError;
-#[cfg(feature = "napi")]
-use crate::ops::{Operation, OutputFormat, PresetConfig, ResizeFit};
-#[cfg(not(feature = "napi"))]
-use crate::ops::{Operation, PresetConfig};
-use image::DynamicImage;
-#[cfg(feature = "napi")]
-use image::GenericImageView;
-#[cfg(feature = "napi")]
-use image::ImageReader;
-#[cfg(feature = "napi")]
-use std::io::Cursor;
-#[cfg(feature = "napi")]
-use std::path::PathBuf;
-#[cfg(feature = "napi")]
-use std::str::FromStr;
-use std::sync::Arc;
-use std::sync::OnceLock;
-
-#[cfg(feature = "napi")]
-use napi::bindgen_prelude::*;
-
-#[cfg(feature = "napi")]
-#[derive(Default)]
-#[napi(object)]
-pub struct KeepMetadataOptions {
-    /// Preserve ICC profile (default: true when options are provided)
-    pub icc: Option<bool>,
-    /// Preserve EXIF metadata (camera info, settings, etc.)
-    /// Note: Orientation tag is automatically reset to 1 after auto-orient
-    pub exif: Option<bool>,
-    /// XMP preservation is not implemented yet (will emit runtime warning when requested)
-    pub xmp: Option<bool>,
-    /// Strip GPS/location tags from EXIF (default: true for privacy protection)
-    /// This is a security-first default that exceeds Sharp's capabilities
-    pub strip_gps: Option<bool>,
-}
-
-#[cfg(feature = "napi")]
-#[derive(Default)]
-#[napi(object)]
-pub struct ResizeOptions {
-    /// Target width in pixels
-    pub width: Option<f64>,
-    /// Target height in pixels
-    pub height: Option<f64>,
-    /// Resize fit mode: inside (default), cover, fill
-    pub fit: Option<String>,
-}
-
-#[cfg(feature = "napi")]
-fn napi_err(env: &Env, err: LazyImageError) -> napi::Error {
-    // Helper to attach code/category consistently when Env is available
-    crate::error::napi_error_with_code(env, err).expect("failed to create napi error")
-}
-
-// Validation helpers are in the sibling `validation` module (src/engine/validation.rs).
-// Re-import under the `validation` namespace so all call-sites remain unchanged.
-#[cfg(feature = "napi")]
-use super::validation;
-
-// MetadataPolicy lives in the sibling `metadata` module (src/engine/metadata.rs).
-// Re-export so downstream code (tasks.rs, engine.rs tests) can still reach it
-// via `super::api::MetadataPolicy` or `crate::engine::api::MetadataPolicy`.
-pub use super::metadata::MetadataPolicy;
-
-/// The main image processing engine.
-///
-/// Usage:
-/// ```js
-/// const result = await ImageEngine.from(buffer)
-///   .resize(800)
-///   .rotate(90)
-///   .grayscale()
-///   .toBuffer('jpeg', 75);
-/// ```
-#[cfg_attr(feature = "napi", napi)]
-pub struct ImageEngine {
-    /// Image source - supports in-memory data and memory-mapped files
-    pub(crate) source: Option<Source>,
-    /// Decoded image (populated after first decode or on sync operations)
-    /// Uses Arc to share decoded image between engines. Combined with Cow<DynamicImage>
-    /// in apply_ops, this enables true Copy-on-Write: no deep copy until mutation.
-    pub(crate) decoded: Option<Arc<DynamicImage>>,
-    /// Queued operations
-    pub(crate) ops: Vec<Operation>,
-    /// Last applied preset (used by toPresetBuffer/toPresetFile convenience APIs)
-    pub(crate) last_preset: Option<PresetConfig>,
-    /// ICC color profile extracted from source image (lazy: extracted on first access)
-    pub(crate) icc_profile: OnceLock<Option<Arc<Vec<u8>>>>,
-    /// Raw EXIF data extracted from source image (lazy: extracted on first access)
-    /// Stored as raw bytes to avoid serialization issues with little_exif::Metadata
-    pub(crate) exif_data: OnceLock<Option<Arc<Vec<u8>>>>,
-    /// Whether to auto-apply EXIF Orientation (default: true)
-    pub(crate) auto_orient: bool,
-    /// Single source of truth for metadata preservation decisions.
-    pub(crate) metadata_policy: MetadataPolicy,
-    /// Whether we already emitted a warning about XMP being unsupported.
-    pub(crate) xmp_warning_emitted: bool,
-    pub(crate) firewall: FirewallConfig,
-}
-
-#[cfg(feature = "napi")]
-#[napi]
-impl ImageEngine {
-    // =========================================================================
-    // CONSTRUCTORS
-    // =========================================================================
-
-    /// Create engine from a buffer.
-    /// Full pixel decode, queued operations, and metadata extraction are all
-    /// deferred until output methods run (true lazy evaluation).
-    #[napi(factory)]
-    pub fn from(buffer: Buffer) -> Self {
-        let data = buffer.to_vec();
-
-        ImageEngine {
-            source: Some(Source::from_vec(data)),
-            decoded: None,
-            ops: Vec::new(),
-            last_preset: None,
-            icc_profile: OnceLock::new(),
-            exif_data: OnceLock::new(),
-            auto_orient: true,
-            metadata_policy: MetadataPolicy::default_policy(),
-            xmp_warning_emitted: false,
-            firewall: FirewallConfig::disabled(),
-        }
-    }
-
-    /// Create engine from a file path.
-    /// Full pixel decode, queued operations, and metadata extraction are all
-    /// deferred until output methods run (true lazy evaluation).
-    /// Small/medium files are read into memory to prevent SIGBUS; very large files (>256 MB)
-    /// use mmap with advisory locks.
-    /// This is the recommended way for server-side processing of large images.
-    #[napi(factory, js_name = "fromPath")]
-    pub fn from_path(env: Env, path: String) -> Result<Self> {
-        if path.trim().is_empty() {
-            return Err(napi_err(
-                &env,
-                LazyImageError::invalid_argument("path", "<empty>", "path must not be empty"),
-            ));
-        }
-
-        let path_buf = PathBuf::from(&path);
-
-        let source = match load_file_safe(&path_buf) {
-            Ok(s) => s,
-            Err(e) => return Err(crate::error::napi_error_with_code(&env, e)?),
-        };
-
-        Ok(ImageEngine {
-            source: Some(source),
-            decoded: None,
-            ops: Vec::new(),
-            last_preset: None,
-            icc_profile: OnceLock::new(),
-            exif_data: OnceLock::new(),
-            auto_orient: true,
-            metadata_policy: MetadataPolicy::default_policy(),
-            xmp_warning_emitted: false,
-            firewall: FirewallConfig::disabled(),
-        })
-    }
-
-    /// Create a clone of this engine (for multi-output scenarios)
-    #[napi(js_name = "clone")]
-    pub fn clone_engine(&self) -> Result<ImageEngine> {
-        // If OnceLock is already initialized, clone the value into a new initialized OnceLock.
-        // If not yet initialized, create a fresh uninitialized OnceLock (avoids forcing extraction).
-        let icc_profile = {
-            let lock = OnceLock::new();
-            if let Some(val) = self.icc_profile.get() {
-                let _ = lock.set(val.clone());
-            }
-            lock
-        };
-        let exif_data = {
-            let lock = OnceLock::new();
-            if let Some(val) = self.exif_data.get() {
-                let _ = lock.set(val.clone());
-            }
-            lock
-        };
-
-        Ok(ImageEngine {
-            source: self.source.clone(),
-            decoded: self.decoded.clone(),
-            ops: self.ops.clone(),
-            last_preset: self.last_preset.clone(),
-            icc_profile,
-            exif_data,
-            auto_orient: self.auto_orient,
-            metadata_policy: self.metadata_policy.clone(),
-            xmp_warning_emitted: self.xmp_warning_emitted,
-            firewall: self.firewall.clone(),
-        })
-    }
-
-    // =========================================================================
-    // PIPELINE OPERATIONS - All return Reference for JS method chaining
-    // =========================================================================
-
-    /// Resize image. Width or height can be null to maintain aspect ratio.
-    /// When both width and height are provided, `fit` controls behavior:
-    /// - "inside" (default): maintain aspect ratio and fit within the box
-    /// - "cover": maintain aspect ratio and crop to fill the box
-    /// - "fill": ignore aspect ratio and force exact dimensions
-    ///
-    /// Supports both signatures:
-    /// - resize({ width?, height?, fit? })
-    /// - resize(width?, height?, fit?)
-    #[napi]
-    pub fn resize(
-        &mut self,
-        env: Env,
-        this: Reference<ImageEngine>,
-        options_or_width: Either<ResizeOptions, Option<f64>>,
-        height: Option<f64>,
-        fit: Option<String>,
-    ) -> Result<Reference<ImageEngine>> {
-        let (width, height, fit) = match options_or_width {
-            Either::A(options) => (options.width, options.height, options.fit),
-            Either::B(width) => (width, height, fit),
-        };
-
-        let fit_mode = if let Some(value) = fit {
-            ResizeFit::from_str(&value)
-                .map_err(|_| napi_err(&env, LazyImageError::invalid_resize_fit(value)))?
-        } else {
-            ResizeFit::default()
-        };
-
-        let (width, height) =
-            validation::sanitize_resize_dimensions(width, height).map_err(|e| napi_err(&env, e))?;
-
-        self.ops.push(Operation::Resize {
-            width,
-            height,
-            fit: fit_mode,
-        });
-        Ok(this)
-    }
-
-    /// Crop a region from the image.
-    #[napi]
-    pub fn crop(
-        &mut self,
-        env: Env,
-        this: Reference<ImageEngine>,
-        x: f64,
-        y: f64,
-        width: f64,
-        height: f64,
-    ) -> Result<Reference<ImageEngine>> {
-        let (x, y, width, height) =
-            validation::sanitize_crop(x, y, width, height).map_err(|e| napi_err(&env, e))?;
-        self.ops.push(Operation::Crop {
-            x,
-            y,
-            width,
-            height,
-        });
-        Ok(this)
-    }
-
-    /// Rotate by degrees (90, 180, 270 only)
-    #[napi]
-    pub fn rotate(
-        &mut self,
-        env: Env,
-        this: Reference<ImageEngine>,
-        degrees: i32,
-    ) -> Result<Reference<ImageEngine>> {
-        match degrees {
-            0 | 90 | 180 | 270 | -90 | -180 | -270 => {
-                self.ops.push(Operation::Rotate { degrees });
-                Ok(this)
-            }
-            other => Err(napi_err(
-                &env,
-                LazyImageError::invalid_rotation_angle(other),
-            )),
-        }
-    }
-
-    /// Flip horizontally
-    #[napi(js_name = "flipH")]
-    pub fn flip_h(&mut self, this: Reference<ImageEngine>) -> Reference<ImageEngine> {
-        self.ops.push(Operation::FlipH);
-        this
-    }
-
-    /// Flip vertically
-    #[napi(js_name = "flipV")]
-    pub fn flip_v(&mut self, this: Reference<ImageEngine>) -> Reference<ImageEngine> {
-        self.ops.push(Operation::FlipV);
-        this
-    }
-
-    /// Convert to grayscale
-    #[napi]
-    pub fn grayscale(&mut self, this: Reference<ImageEngine>) -> Reference<ImageEngine> {
-        self.ops.push(Operation::Grayscale);
-        this
-    }
-
-    /// Enable or disable EXIF auto-orientation (default: enabled).
-    /// `true` = apply EXIF Orientation automatically (sharp-compatible)
-    /// `false` = ignore EXIF Orientation
-    #[napi(js_name = "autoOrient")]
-    pub fn auto_orient(
-        &mut self,
-        this: Reference<ImageEngine>,
-        enabled: bool,
-    ) -> Reference<ImageEngine> {
-        self.auto_orient = enabled;
-        this
-    }
-
-    /// Preserve metadata in output.
-    /// - ICC profile: Preserved when `icc: true` (default when options provided)
-    /// - EXIF: Preserved when `exif: true`. Orientation is auto-reset to 1 after auto-orient.
-    /// - GPS: Stripped by default when `stripGps: true` (privacy-first, exceeds Sharp)
-    /// - XMP: Not yet supported (emits warning)
-    ///
-    /// By default, all metadata is stripped for security and smaller file sizes.
-    #[napi(js_name = "keepMetadata")]
-    pub fn keep_metadata(
-        &mut self,
-        _env: Env,
-        this: Reference<ImageEngine>,
-        options: Option<KeepMetadataOptions>,
-    ) -> Result<Reference<ImageEngine>> {
-        let opts = options.unwrap_or_default();
-        let icc = opts.icc.unwrap_or(true);
-        let exif = opts.exif.unwrap_or(false);
-        let xmp = opts.xmp.unwrap_or(false);
-        let strip_gps = opts.strip_gps.unwrap_or(true); // Default: strip GPS for privacy
-
-        // XMP is not yet supported - emit warning once
-        if xmp && !self.xmp_warning_emitted {
-            let warning =
-                "keepMetadata: XMP preservation is not implemented yet; XMP data will be stripped.";
-            eprintln!("{}", warning);
-            self.xmp_warning_emitted = true;
-        }
-
-        self.metadata_policy.icc = icc;
-        self.metadata_policy.exif = exif;
-        self.metadata_policy.gps = !strip_gps; // gps=true means keep, strip_gps=true means remove
-        Ok(this)
-    }
-
-    /// Enable Image Firewall mode with built-in policies (strict or lenient).
-    /// Strict mode enforces aggressive limits and rejects dangerous metadata (best for zero-trust inputs).
-    /// Lenient mode keeps generous limits but still guards against decompression bombs.
-    #[napi]
-    pub fn sanitize(
-        &mut self,
-        env: Env,
-        this: Reference<ImageEngine>,
-        options: Option<SanitizeOptions>,
-    ) -> Result<Reference<ImageEngine>> {
-        let requested = options
-            .and_then(|opts| opts.policy)
-            .unwrap_or_else(|| "strict".to_string());
-        let lowered = requested.to_lowercase();
-        let policy = match lowered.as_str() {
-            "strict" => FirewallPolicy::Strict,
-            "lenient" => FirewallPolicy::Lenient,
-            _ => {
-                return Err(napi_err(
-                    &env,
-                    LazyImageError::invalid_firewall_policy(requested),
-                ))
-            }
-        };
-
-        self.firewall = FirewallConfig::apply_policy(policy);
-        if self.firewall.reject_metadata {
-            self.metadata_policy.apply_firewall(true);
-            // No need to clear the OnceLock fields: output methods already
-            // check metadata_policy and skip metadata when firewall is active.
-        }
-
-        if let Some(decoded) = &self.decoded {
-            self.firewall
-                .enforce_pixels(decoded.width(), decoded.height())
-                .map_err(|e| napi_err(&env, e))?;
-        }
-
-        if let Some(source) = &self.source {
-            self.firewall
-                .enforce_source_len(source.len())
-                .map_err(|e| napi_err(&env, e))?;
-            if let Some(bytes) = source.as_bytes() {
-                self.firewall
-                    .scan_metadata(bytes)
-                    .map_err(|e| napi_err(&env, e))?;
-
-                // When the image is not yet decoded (lazy path, e.g. fromPath),
-                // read the header to check dimensions against the firewall policy.
-                // This ensures sanitize() rejects oversized images immediately for
-                // both from(buffer) and fromPath(path) construction paths.
-                if self.decoded.is_none() {
-                    let cursor = Cursor::new(bytes);
-                    if let Ok(reader) = ImageReader::new(cursor).with_guessed_format() {
-                        if let Ok((w, h)) = reader.into_dimensions() {
-                            self.firewall
-                                .enforce_pixels(w, h)
-                                .map_err(|e| napi_err(&env, e))?;
-                        }
-                    }
-                }
-            }
-        }
-
-        Ok(this)
-    }
-
-    /// Override Image Firewall limits (maxPixels, maxBytes, timeoutMs).
-    /// Any field set to 0 disables that particular limit.
-    #[napi]
-    pub fn limits(
-        &mut self,
-        env: Env,
-        this: Reference<ImageEngine>,
-        options: FirewallLimitOptions,
-    ) -> Result<Reference<ImageEngine>> {
-        if !self.firewall.enabled || matches!(self.firewall.policy, FirewallPolicy::Disabled) {
-            self.firewall = FirewallConfig::custom();
-        } else {
-            self.firewall.enabled = true;
-            self.firewall.policy = FirewallPolicy::Custom;
-        }
-
-        match validation::sanitize_limit("maxPixels", options.max_pixels)
-            .map_err(|e| napi_err(&env, e))?
-        {
-            validation::LimitUpdate::Unchanged => {}
-            validation::LimitUpdate::Disable => self.firewall.max_pixels = None,
-            validation::LimitUpdate::Set(value) => self.firewall.max_pixels = Some(value),
-        }
-
-        match validation::sanitize_limit("maxBytes", options.max_bytes)
-            .map_err(|e| napi_err(&env, e))?
-        {
-            validation::LimitUpdate::Unchanged => {}
-            validation::LimitUpdate::Disable => self.firewall.max_bytes = None,
-            validation::LimitUpdate::Set(value) => self.firewall.max_bytes = Some(value),
-        }
-
-        match validation::sanitize_limit("timeoutMs", options.timeout_ms)
-            .map_err(|e| napi_err(&env, e))?
-        {
-            validation::LimitUpdate::Unchanged => {}
-            validation::LimitUpdate::Disable => self.firewall.timeout_ms = None,
-            validation::LimitUpdate::Set(value) => self.firewall.timeout_ms = Some(value),
-        }
-
-        Ok(this)
-    }
-
-    /// Adjust brightness (-100 to 100)
-    #[napi]
-    pub fn brightness(
-        &mut self,
-        env: Env,
-        this: Reference<ImageEngine>,
-        value: f64,
-    ) -> Result<Reference<ImageEngine>> {
-        let value = validation::ensure_finite_integer("brightness", value)
-            .and_then(|int| {
-                if !(-100..=100).contains(&int) {
-                    Err(LazyImageError::invalid_argument(
-                        "brightness",
-                        int.to_string(),
-                        "expected value between -100 and 100",
-                    ))
-                } else {
-                    Ok(int as i32)
-                }
-            })
-            .map_err(|e| napi_err(&env, e))?;
-
-        self.ops.push(Operation::Brightness { value });
-        Ok(this)
-    }
-
-    /// Adjust contrast (-100 to 100)
-    #[napi]
-    pub fn contrast(
-        &mut self,
-        env: Env,
-        this: Reference<ImageEngine>,
-        value: f64,
-    ) -> Result<Reference<ImageEngine>> {
-        let value = validation::ensure_finite_integer("contrast", value)
-            .and_then(|int| {
-                if !(-100..=100).contains(&int) {
-                    Err(LazyImageError::invalid_argument(
-                        "contrast",
-                        int.to_string(),
-                        "expected value between -100 and 100",
-                    ))
-                } else {
-                    Ok(int as i32)
-                }
-            })
-            .map_err(|e| napi_err(&env, e))?;
-
-        self.ops.push(Operation::Contrast { value });
-        Ok(this)
-    }
-
-    /// Normalize pixel format to RGB/RGBA without performing any color space transformation.
-    /// This does not apply ICC profile conversion; it only guarantees the pixel layout is RGB/RGBA.
-    /// Use a dedicated color management library for true color space conversions.
-    #[napi(js_name = "normalizePixelFormat")]
-    pub fn normalize_pixel_format(
-        &mut self,
-        this: Reference<ImageEngine>,
-    ) -> Result<Reference<ImageEngine>> {
-        // Normalize to RGB/RGBA pixel format (no color space conversion or ICC processing)
-        self.ops.push(Operation::ColorSpace {
-            target: crate::ops::ColorSpace::Srgb,
-        });
-        Ok(this)
-    }
-
-    // =========================================================================
-    // PRESETS - Common configurations for web optimization
-    // =========================================================================
-
-    /// Apply a built-in preset for common use cases.
-    ///
-    /// Available presets:
-    /// - "thumbnail": 150x150, WebP quality 75 (gallery thumbnails)
-    /// - "avatar": 200x200, WebP quality 80 (profile pictures)
-    /// - "hero": 1920 width, JPEG quality 85 (hero images, banners)
-    /// - "social": 1200x630, JPEG quality 80 (OGP/Twitter cards)
-    ///
-    /// Returns the preset configuration for use with toBuffer/toFile.
-    #[napi]
-    pub fn preset(
-        &mut self,
-        env: Env,
-        _this: Reference<ImageEngine>,
-        name: String,
-    ) -> Result<PresetResult> {
-        let config = match PresetConfig::get(&name) {
-            Some(config) => config,
-            None => {
-                let lazy_err = LazyImageError::invalid_preset(name.clone());
-                return Err(crate::error::napi_error_with_code(&env, lazy_err)?);
-            }
-        };
-
-        // Apply resize operation
-        self.ops.push(Operation::Resize {
-            width: config.width,
-            height: config.height,
-            fit: ResizeFit::Inside,
-        });
-
-        // Return preset info for the user to use with toBuffer/toFile
-        let (format_str, quality) = match &config.format {
-            OutputFormat::Jpeg { quality, .. } => ("jpeg", Some(*quality)),
-            OutputFormat::Png => ("png", None),
-            OutputFormat::WebP { quality } => ("webp", Some(*quality)),
-            OutputFormat::Avif { quality } => ("avif", Some(*quality)),
-        };
-
-        // Store last preset for convenience helpers
-        self.last_preset = Some(config.clone());
-
-        Ok(PresetResult {
-            format: format_str.to_string(),
-            quality,
-            width: config.width,
-            height: config.height,
-        })
-    }
-
-    // =========================================================================
-    // OUTPUT - Triggers async computation
-    // =========================================================================
-
-    /// Encode to buffer asynchronously.
-    /// format: "jpeg", "jpg", "png", "webp", "avif"
-    /// quality: 1-100 for lossy formats (default: JPEG=85, WebP=80, AVIF=60). Omit for PNG.
-    /// fastMode: If true, uses faster encoding for JPEG (2-4x faster, slightly larger files). Default: false.
-    ///
-    /// **Non-destructive**: This method can be called multiple times on the same engine instance.
-    /// The source data is cloned internally, allowing multiple format outputs.
-    #[napi(ts_return_type = "Promise<Buffer>")]
-    pub fn to_buffer(
-        &mut self,
-        env: Env,
-        format: String,
-        quality: Option<f64>,
-        fast_mode: Option<bool>,
-    ) -> Result<AsyncTask<EncodeTask>> {
-        let fast_mode = fast_mode.unwrap_or(false);
-        let quality = validation::sanitize_quality(quality).map_err(|e| napi_err(&env, e))?;
-        let output_format = match OutputFormat::from_str_with_options(&format, quality, fast_mode) {
-            Ok(format) => format,
-            Err(_e) => {
-                let lazy_err = LazyImageError::unsupported_format(format.clone());
-                return Err(crate::error::napi_error_with_code(&env, lazy_err)?);
-            }
-        };
-
-        // Use source directly - zero-copy for Memory and Mapped sources
-        let source = self.source.clone();
-        let decoded = self.decoded.clone();
-        let ops = self.ops.clone();
-        let mut policy = self.metadata_policy.clone();
-        policy.apply_firewall(self.firewall.reject_metadata);
-        let auto_orient = self.auto_orient;
-        let icc_present = self.icc_profile().is_some();
-        let icc_profile = if policy.effective_icc() {
-            self.icc_profile().cloned()
-        } else {
-            None
-        };
-        let exif_data = if policy.effective_exif() {
-            self.exif_data().cloned()
-        } else {
-            None
-        };
-
-        Ok(AsyncTask::new(EncodeTask {
-            ctx: TaskContext {
-                source,
-                decoded,
-                ops,
-                format: output_format,
-                icc_profile,
-                icc_present,
-                exif_data,
-                auto_orient,
-                metadata_policy: policy,
-                firewall: self.firewall.clone(),
-                #[cfg(feature = "napi")]
-                last_error: None,
-            },
-        }))
-    }
-
-    /// Convenience: encode using the last applied preset by name.
-    /// Equivalent to calling `preset(name)` then `toBuffer(preset.format, preset.quality)`.
-    #[napi(js_name = "toBufferWithPreset", ts_return_type = "Promise<Buffer>")]
-    pub fn to_buffer_with_preset(
-        &mut self,
-        env: Env,
-        preset_name: String,
-    ) -> Result<AsyncTask<EncodeTask>> {
-        let preset = match PresetConfig::get(&preset_name) {
-            Some(config) => config,
-            None => {
-                let lazy_err = LazyImageError::invalid_preset(preset_name.clone());
-                return Err(crate::error::napi_error_with_code(&env, lazy_err)?);
-            }
-        };
-
-        // Apply resize ops in-place, mirroring preset()
-        self.ops.push(Operation::Resize {
-            width: preset.width,
-            height: preset.height,
-            fit: ResizeFit::Inside,
-        });
-
-        self.last_preset = Some(preset.clone());
-
-        let (format_str, quality, fast_mode) = match &preset.format {
-            OutputFormat::Jpeg { quality, fast_mode } => ("jpeg", Some(*quality), Some(*fast_mode)),
-            OutputFormat::Png => ("png", None, None),
-            OutputFormat::WebP { quality } => ("webp", Some(*quality), None),
-            OutputFormat::Avif { quality } => ("avif", Some(*quality), None),
-        };
-
-        self.to_buffer(
-            env,
-            format_str.to_string(),
-            quality.map(|q| q as f64),
-            fast_mode,
-        )
-    }
-
-    /// Encode to buffer asynchronously with performance metrics.
-    /// Returns `{ data: Buffer, metrics: ProcessingMetrics }`.
-    ///
-    /// **Non-destructive**: This method can be called multiple times on the same engine instance.
-    /// The source data is cloned internally, allowing multiple format outputs.
-    #[napi(ts_return_type = "Promise<OutputWithMetrics>")]
-    pub fn to_buffer_with_metrics(
-        &mut self,
-        env: Env,
-        format: String,
-        quality: Option<f64>,
-        fast_mode: Option<bool>,
-    ) -> Result<AsyncTask<EncodeWithMetricsTask>> {
-        let fast_mode = fast_mode.unwrap_or(false);
-        let quality = validation::sanitize_quality(quality).map_err(|e| napi_err(&env, e))?;
-        let output_format = match OutputFormat::from_str_with_options(&format, quality, fast_mode) {
-            Ok(format) => format,
-            Err(_e) => {
-                let lazy_err = LazyImageError::unsupported_format(format.clone());
-                return Err(crate::error::napi_error_with_code(&env, lazy_err)?);
-            }
-        };
-
-        // Use source directly - zero-copy for Memory and Mapped sources
-        let source = self.source.clone();
-        let decoded = self.decoded.clone();
-        let ops = self.ops.clone();
-        let mut policy = self.metadata_policy.clone();
-        policy.apply_firewall(self.firewall.reject_metadata);
-        let auto_orient = self.auto_orient;
-        let icc_present = self.icc_profile().is_some();
-        let icc_profile = if policy.effective_icc() {
-            self.icc_profile().cloned()
-        } else {
-            None
-        };
-        let exif_data = if policy.effective_exif() {
-            self.exif_data().cloned()
-        } else {
-            None
-        };
-
-        Ok(AsyncTask::new(EncodeWithMetricsTask {
-            ctx: TaskContext {
-                source,
-                decoded,
-                ops,
-                format: output_format,
-                icc_profile,
-                icc_present,
-                exif_data,
-                auto_orient,
-                metadata_policy: policy,
-                firewall: self.firewall.clone(),
-                #[cfg(feature = "napi")]
-                last_error: None,
-            },
-        }))
-    }
-
-    /// Convenience: encode with metrics using a preset name.
-    /// Equivalent to `preset(name)` then `toBufferWithMetrics(preset.format, preset.quality)`.
-    #[napi(
-        js_name = "toBufferWithMetricsPreset",
-        ts_return_type = "Promise<OutputWithMetrics>"
-    )]
-    pub fn to_buffer_with_metrics_preset(
-        &mut self,
-        env: Env,
-        preset_name: String,
-    ) -> Result<AsyncTask<EncodeWithMetricsTask>> {
-        let preset = match PresetConfig::get(&preset_name) {
-            Some(config) => config,
-            None => {
-                let lazy_err = LazyImageError::invalid_preset(preset_name.clone());
-                return Err(crate::error::napi_error_with_code(&env, lazy_err)?);
-            }
-        };
-
-        self.ops.push(Operation::Resize {
-            width: preset.width,
-            height: preset.height,
-            fit: ResizeFit::Inside,
-        });
-
-        self.last_preset = Some(preset.clone());
-
-        let (format_str, quality, fast_mode) = match &preset.format {
-            OutputFormat::Jpeg { quality, fast_mode } => ("jpeg", Some(*quality), Some(*fast_mode)),
-            OutputFormat::Png => ("png", None, None),
-            OutputFormat::WebP { quality } => ("webp", Some(*quality), None),
-            OutputFormat::Avif { quality } => ("avif", Some(*quality), None),
-        };
-
-        self.to_buffer_with_metrics(
-            env,
-            format_str.to_string(),
-            quality.map(|q| q as f64),
-            fast_mode,
-        )
-    }
-
-    /// Encode to buffer with a byte-budget constraint.
-    ///
-    /// Performs a binary search over quality (entirely in Rust) to find the
-    /// highest quality that keeps the output within `target_bytes`. This
-    /// eliminates ~7 JS↔NAPI round-trips per call compared to the previous
-    /// JS-side implementation.
-    ///
-    /// Returns `TargetBytesResult` with the encoded data, chosen quality,
-    /// actual size, and whether the budget was met.
-    #[napi(
-        js_name = "toBufferTargetBytesNative",
-        ts_return_type = "Promise<TargetBytesResult>"
-    )]
-    pub fn to_buffer_target_bytes_native(
-        &mut self,
-        env: Env,
-        format: String,
-        target_bytes: u32,
-        min_quality: Option<u32>,
-        max_quality: Option<u32>,
-        fast_mode: Option<bool>,
-        strict: Option<bool>,
-    ) -> Result<AsyncTask<EncodeTargetBytesTask>> {
-        let fast_mode = fast_mode.unwrap_or(false);
-
-        // Validate and clamp quality range
-        let min_q = min_quality.unwrap_or(30).min(100) as u8;
-        let max_q = max_quality.unwrap_or(100).min(100) as u8;
-        if min_q > max_q {
-            return Err(napi_err(
-                &env,
-                LazyImageError::invalid_argument(
-                    "minQuality",
-                    min_q.to_string(),
-                    "must be less than or equal to maxQuality",
-                ),
-            ));
-        }
-        if min_q == 0 {
-            return Err(napi_err(
-                &env,
-                LazyImageError::invalid_argument("minQuality", "0", "must be between 1 and 100"),
-            ));
-        }
-        if target_bytes == 0 {
-            return Err(napi_err(
-                &env,
-                LazyImageError::invalid_argument("targetBytes", "0", "must be a positive number"),
-            ));
-        }
-
-        // Parse format (quality is irrelevant here, will be overridden per iteration)
-        let output_format =
-            match OutputFormat::from_str_with_options(&format, Some(max_q), fast_mode) {
-                Ok(f) => f,
-                Err(_) => {
-                    let lazy_err = LazyImageError::unsupported_format(format.clone());
-                    return Err(crate::error::napi_error_with_code(&env, lazy_err)?);
-                }
-            };
-
-        let source = self.source.clone();
-        let decoded = self.decoded.clone();
-        let ops = self.ops.clone();
-        let mut policy = self.metadata_policy.clone();
-        policy.apply_firewall(self.firewall.reject_metadata);
-        let auto_orient = self.auto_orient;
-        let icc_present = self.icc_profile().is_some();
-        let icc_profile = if policy.effective_icc() {
-            self.icc_profile().cloned()
-        } else {
-            None
-        };
-        let exif_data = if policy.effective_exif() {
-            self.exif_data().cloned()
-        } else {
-            None
-        };
-
-        Ok(AsyncTask::new(EncodeTargetBytesTask {
-            ctx: TaskContext {
-                source,
-                decoded,
-                ops,
-                format: output_format,
-                icc_profile,
-                icc_present,
-                exif_data,
-                auto_orient,
-                metadata_policy: policy,
-                firewall: self.firewall.clone(),
-                #[cfg(feature = "napi")]
-                last_error: None,
-            },
-            target_bytes,
-            min_quality: min_q,
-            max_quality: max_q,
-            quality_floor_policy_strict: strict.unwrap_or(false),
-        }))
-    }
-
-    /// Encode and write directly to a file asynchronously.
-    /// **Memory-efficient**: Combined with fromPath(), this enables
-    /// full file-to-file processing without touching Node.js heap.
-    ///
-    /// **Non-destructive**: This method can be called multiple times on the same engine instance.
-    /// The source data is cloned internally, allowing multiple format outputs.
-    ///
-    /// Returns the number of bytes written.
-    #[napi(js_name = "toFile", ts_return_type = "Promise<number>")]
-    pub fn to_file(
-        &mut self,
-        env: Env,
-        path: String,
-        format: String,
-        quality: Option<f64>,
-        fast_mode: Option<bool>,
-    ) -> Result<AsyncTask<WriteFileTask>> {
-        let fast_mode = fast_mode.unwrap_or(false);
-        let quality = validation::sanitize_quality(quality).map_err(|e| napi_err(&env, e))?;
-        validation::validate_output_path(&path).map_err(|e| napi_err(&env, e))?;
-
-        let output_format = match OutputFormat::from_str_with_options(&format, quality, fast_mode) {
-            Ok(format) => format,
-            Err(_e) => {
-                let lazy_err = LazyImageError::unsupported_format(format.clone());
-                return Err(crate::error::napi_error_with_code(&env, lazy_err)?);
-            }
-        };
-
-        // Use source directly - zero-copy for Memory and Mapped sources
-        let source = self.source.clone();
-        let decoded = self.decoded.clone();
-        let ops = self.ops.clone();
-        let mut policy = self.metadata_policy.clone();
-        policy.apply_firewall(self.firewall.reject_metadata);
-        let auto_orient = self.auto_orient;
-        let icc_present = self.icc_profile().is_some();
-        let icc_profile = if policy.effective_icc() {
-            self.icc_profile().cloned()
-        } else {
-            None
-        };
-        let exif_data = if policy.effective_exif() {
-            self.exif_data().cloned()
-        } else {
-            None
-        };
-
-        Ok(AsyncTask::new(WriteFileTask {
-            ctx: TaskContext {
-                source,
-                decoded,
-                ops,
-                format: output_format,
-                icc_profile,
-                icc_present,
-                exif_data,
-                auto_orient,
-                metadata_policy: policy,
-                firewall: self.firewall.clone(),
-                #[cfg(feature = "napi")]
-                last_error: None,
-            },
-            output_path: path,
-        }))
-    }
-
-    /// Encode and write directly to a file asynchronously, returning metrics.
-    #[napi(
-        js_name = "toFileWithMetrics",
-        ts_return_type = "Promise<FileOutputWithMetrics>"
-    )]
-    pub fn to_file_with_metrics(
-        &mut self,
-        env: Env,
-        path: String,
-        format: String,
-        quality: Option<f64>,
-        fast_mode: Option<bool>,
-    ) -> Result<AsyncTask<WriteFileWithMetricsTask>> {
-        let fast_mode = fast_mode.unwrap_or(false);
-        let quality = validation::sanitize_quality(quality).map_err(|e| napi_err(&env, e))?;
-        validation::validate_output_path(&path).map_err(|e| napi_err(&env, e))?;
-
-        let output_format = match OutputFormat::from_str_with_options(&format, quality, fast_mode) {
-            Ok(format) => format,
-            Err(_e) => {
-                let lazy_err = LazyImageError::unsupported_format(format.clone());
-                return Err(crate::error::napi_error_with_code(&env, lazy_err)?);
-            }
-        };
-
-        let source = self.source.clone();
-        let decoded = self.decoded.clone();
-        let ops = self.ops.clone();
-        let mut policy = self.metadata_policy.clone();
-        policy.apply_firewall(self.firewall.reject_metadata);
-        let auto_orient = self.auto_orient;
-        let icc_present = self.icc_profile().is_some();
-        let icc_profile = if policy.effective_icc() {
-            self.icc_profile().cloned()
-        } else {
-            None
-        };
-        let exif_data = if policy.effective_exif() {
-            self.exif_data().cloned()
-        } else {
-            None
-        };
-
-        Ok(AsyncTask::new(WriteFileWithMetricsTask {
-            ctx: TaskContext {
-                source,
-                decoded,
-                ops,
-                format: output_format,
-                icc_profile,
-                icc_present,
-                exif_data,
-                auto_orient,
-                metadata_policy: policy,
-                firewall: self.firewall.clone(),
-                #[cfg(feature = "napi")]
-                last_error: None,
-            },
-            output_path: path,
-        }))
-    }
-
-    // =========================================================================
-    // UNIFIED OUTPUT API — encode() and encodeToFile()
-    //
-    // These two methods consolidate all the individual output methods above
-    // into a single options-object API.  Existing methods remain for backward
-    // compatibility; new code should prefer encode() / encodeToFile().
-    // =========================================================================
-
-    /// Unified encode-to-buffer method.
-    ///
-    /// Accepts an options object with `format`, `quality`, `fastMode`, `preset`,
-    /// and `metrics` fields.  When `preset` is supplied the preset's settings
-    /// take precedence over format/quality/fastMode.
-    ///
-    /// Returns `EncodeResult { data, metrics? }`.
-    #[napi(ts_return_type = "Promise<EncodeResult>")]
-    pub fn encode(
-        &mut self,
-        env: Env,
-        options: crate::EncodeOptions,
-    ) -> Result<AsyncTask<crate::engine::tasks::UnifiedEncodeTask>> {
-        let want_metrics = options.metrics.unwrap_or(false);
-
-        // Resolve format/quality/fastMode — preset wins when supplied.
-        let (format_str, quality_f64, fast_mode) = if let Some(ref preset_name) = options.preset {
-            let preset = match PresetConfig::get(preset_name) {
-                Some(c) => c,
-                None => {
-                    let lazy_err = LazyImageError::invalid_preset(preset_name.clone());
-                    return Err(crate::error::napi_error_with_code(&env, lazy_err)?);
-                }
-            };
-
-            // Apply the preset's resize ops.
-            self.ops.push(Operation::Resize {
-                width: preset.width,
-                height: preset.height,
-                fit: ResizeFit::Inside,
-            });
-            self.last_preset = Some(preset.clone());
-
-            let (f, q, fm) = match &preset.format {
-                OutputFormat::Jpeg { quality, fast_mode } => {
-                    ("jpeg".to_string(), Some(*quality as f64), Some(*fast_mode))
-                }
-                OutputFormat::Png => ("png".to_string(), None, None),
-                OutputFormat::WebP { quality } => ("webp".to_string(), Some(*quality as f64), None),
-                OutputFormat::Avif { quality } => ("avif".to_string(), Some(*quality as f64), None),
-            };
-            (f, q, fm.unwrap_or(false))
-        } else {
-            let fmt = options.format.unwrap_or_else(|| "jpeg".to_string());
-            (fmt, options.quality, options.fast_mode.unwrap_or(false))
-        };
-
-        let quality = validation::sanitize_quality(quality_f64).map_err(|e| napi_err(&env, e))?;
-        let output_format =
-            match OutputFormat::from_str_with_options(&format_str, quality, fast_mode) {
-                Ok(f) => f,
-                Err(_) => {
-                    let lazy_err = LazyImageError::unsupported_format(format_str.clone());
-                    return Err(crate::error::napi_error_with_code(&env, lazy_err)?);
-                }
-            };
-
-        let source = self.source.clone();
-        let decoded = self.decoded.clone();
-        let ops = self.ops.clone();
-        let mut policy = self.metadata_policy.clone();
-        policy.apply_firewall(self.firewall.reject_metadata);
-        let auto_orient = self.auto_orient;
-        let icc_present = self.icc_profile().is_some();
-        let icc_profile = if policy.effective_icc() {
-            self.icc_profile().cloned()
-        } else {
-            None
-        };
-        let exif_data = if policy.effective_exif() {
-            self.exif_data().cloned()
-        } else {
-            None
-        };
-
-        Ok(AsyncTask::new(crate::engine::tasks::UnifiedEncodeTask {
-            ctx: TaskContext {
-                source,
-                decoded,
-                ops,
-                format: output_format,
-                icc_profile,
-                icc_present,
-                exif_data,
-                auto_orient,
-                metadata_policy: policy,
-                firewall: self.firewall.clone(),
-                #[cfg(feature = "napi")]
-                last_error: None,
-            },
-            want_metrics,
-        }))
-    }
-
-    /// Unified encode-to-file method.
-    ///
-    /// Same options as `encode()`, but writes the result directly to `path`.
-    /// Returns `FileEncodeResult { bytesWritten, metrics? }`.
-    #[napi(js_name = "encodeToFile", ts_return_type = "Promise<FileEncodeResult>")]
-    pub fn encode_to_file(
-        &mut self,
-        env: Env,
-        path: String,
-        options: crate::EncodeOptions,
-    ) -> Result<AsyncTask<crate::engine::tasks::UnifiedEncodeToFileTask>> {
-        validation::validate_output_path(&path).map_err(|e| napi_err(&env, e))?;
-        let want_metrics = options.metrics.unwrap_or(false);
-
-        let (format_str, quality_f64, fast_mode) = if let Some(ref preset_name) = options.preset {
-            let preset = match PresetConfig::get(preset_name) {
-                Some(c) => c,
-                None => {
-                    let lazy_err = LazyImageError::invalid_preset(preset_name.clone());
-                    return Err(crate::error::napi_error_with_code(&env, lazy_err)?);
-                }
-            };
-
-            self.ops.push(Operation::Resize {
-                width: preset.width,
-                height: preset.height,
-                fit: ResizeFit::Inside,
-            });
-            self.last_preset = Some(preset.clone());
-
-            let (f, q, fm) = match &preset.format {
-                OutputFormat::Jpeg { quality, fast_mode } => {
-                    ("jpeg".to_string(), Some(*quality as f64), Some(*fast_mode))
-                }
-                OutputFormat::Png => ("png".to_string(), None, None),
-                OutputFormat::WebP { quality } => ("webp".to_string(), Some(*quality as f64), None),
-                OutputFormat::Avif { quality } => ("avif".to_string(), Some(*quality as f64), None),
-            };
-            (f, q, fm.unwrap_or(false))
-        } else {
-            let fmt = options.format.unwrap_or_else(|| "jpeg".to_string());
-            (fmt, options.quality, options.fast_mode.unwrap_or(false))
-        };
-
-        let quality = validation::sanitize_quality(quality_f64).map_err(|e| napi_err(&env, e))?;
-        let output_format =
-            match OutputFormat::from_str_with_options(&format_str, quality, fast_mode) {
-                Ok(f) => f,
-                Err(_) => {
-                    let lazy_err = LazyImageError::unsupported_format(format_str.clone());
-                    return Err(crate::error::napi_error_with_code(&env, lazy_err)?);
-                }
-            };
-
-        let source = self.source.clone();
-        let decoded = self.decoded.clone();
-        let ops = self.ops.clone();
-        let mut policy = self.metadata_policy.clone();
-        policy.apply_firewall(self.firewall.reject_metadata);
-        let auto_orient = self.auto_orient;
-        let icc_present = self.icc_profile().is_some();
-        let icc_profile = if policy.effective_icc() {
-            self.icc_profile().cloned()
-        } else {
-            None
-        };
-        let exif_data = if policy.effective_exif() {
-            self.exif_data().cloned()
-        } else {
-            None
-        };
-
-        Ok(AsyncTask::new(
-            crate::engine::tasks::UnifiedEncodeToFileTask {
-                ctx: TaskContext {
-                    source,
-                    decoded,
-                    ops,
-                    format: output_format,
-                    icc_profile,
-                    icc_present,
-                    exif_data,
-                    auto_orient,
-                    metadata_policy: policy,
-                    firewall: self.firewall.clone(),
-                    #[cfg(feature = "napi")]
-                    last_error: None,
-                },
-                want_metrics,
-                output_path: path,
-            },
-        ))
-    }
-
-    /// Convenience: encode to file using the preset's recommended format/quality.
-    #[napi(js_name = "toFileWithPreset", ts_return_type = "Promise<number>")]
-    pub fn to_file_with_preset(
-        &mut self,
-        env: Env,
-        path: String,
-        preset_name: String,
-    ) -> Result<AsyncTask<WriteFileTask>> {
-        let preset = match PresetConfig::get(&preset_name) {
-            Some(config) => config,
-            None => {
-                let lazy_err = LazyImageError::invalid_preset(preset_name.clone());
-                return Err(crate::error::napi_error_with_code(&env, lazy_err)?);
-            }
-        };
-
-        self.ops.push(Operation::Resize {
-            width: preset.width,
-            height: preset.height,
-            fit: ResizeFit::Inside,
-        });
-
-        self.last_preset = Some(preset.clone());
-
-        let (format_str, quality, fast_mode) = match &preset.format {
-            OutputFormat::Jpeg { quality, fast_mode } => ("jpeg", Some(*quality), Some(*fast_mode)),
-            OutputFormat::Png => ("png", None, None),
-            OutputFormat::WebP { quality } => ("webp", Some(*quality), None),
-            OutputFormat::Avif { quality } => ("avif", Some(*quality), None),
-        };
-
-        self.to_file(
-            env,
-            path,
-            format_str.to_string(),
-            quality.map(|q| q as f64),
-            fast_mode,
-        )
-    }
-
-    // =========================================================================
-    // SYNC UTILITIES
-    // =========================================================================
-
-    /// Get image dimensions WITHOUT full decoding (internal method, no Env required).
-    /// For file paths, reads only the header bytes (extremely fast).
-    /// For in-memory buffers, uses header-only parsing.
-    /// This method is public for testing purposes.
-    pub fn dimensions_internal(&mut self) -> std::result::Result<Dimensions, LazyImageError> {
-        // If already decoded, use that
-        if let Some(ref img) = self.decoded {
-            let (w, h) = img.dimensions();
-            return Ok(Dimensions {
-                width: w,
-                height: h,
-            });
-        }
-
-        // Try to read dimensions from header only (no full decode)
-        let source = match self.source.as_ref() {
-            Some(source) => source,
-            None => {
-                return Err(LazyImageError::source_consumed());
-            }
-        };
-
-        self.firewall.enforce_source_len(source.len())?;
-
-        // Use as_bytes() to get zero-copy access for Memory and Mapped sources
-        if let Some(bytes) = source.as_bytes() {
-            self.firewall.scan_metadata(bytes)?;
-            // For in-memory or memory-mapped data, use cursor
-            let cursor = Cursor::new(bytes);
-            let reader = match ImageReader::new(cursor).with_guessed_format() {
-                Ok(reader) => reader,
-                Err(e) => {
-                    let err_msg = format!("failed to read image header: {e}");
-                    return Err(LazyImageError::decode_failed(err_msg));
-                }
-            };
-
-            let (width, height) = match reader.into_dimensions() {
-                Ok(dims) => dims,
-                Err(e) => {
-                    let err_msg = format!("failed to read dimensions: {e}");
-                    return Err(LazyImageError::decode_failed(err_msg));
-                }
-            };
-
-            Ok(Dimensions { width, height })
-        } else {
-            Err(LazyImageError::source_consumed())
-        }
-    }
-
-    /// Get image dimensions WITHOUT full decoding.
-    /// For file paths, reads only the header bytes (extremely fast).
-    /// For in-memory buffers, uses header-only parsing.
-    #[napi]
-    pub fn dimensions(&mut self, env: Env) -> Result<Dimensions> {
-        match self.dimensions_internal() {
-            Ok(dims) => Ok(dims),
-            Err(lazy_err) => {
-                let napi_err = crate::error::napi_error_with_code(&env, lazy_err)?;
-                Err(napi_err)
-            }
-        }
-    }
-
-    /// Check if an ICC color profile was extracted from the source image.
-    /// Returns the profile size in bytes, or null if no profile exists.
-    /// This triggers lazy extraction of the ICC profile if not yet extracted.
-    #[napi(js_name = "hasIccProfile")]
-    pub fn has_icc_profile(&self) -> Option<u32> {
-        self.icc_profile().map(|p| p.len() as u32)
-    }
-
-    /// Process multiple images in parallel with the same operations.
-    ///
-    /// - inputs: Array of input file paths
-    /// - output_dir: Directory to write processed images
-    /// - options: Output settings
-    ///   - format: Output format ("jpeg", "png", "webp", "avif")
-    ///   - quality: Optional quality (1-100, uses format-specific default if omitted; omit for PNG)
-    ///   - fastMode: Optional fast mode flag (only applies to JPEG, default: false)
-    ///   - concurrency: Optional number of parallel workers:
-    ///     - 0 or undefined: Auto-detect based on CPU cores and memory limits (smart concurrency)
-    ///       Detects container memory limits (cgroup v1/v2) and adjusts to prevent OOM kills.
-    ///       Ideal for serverless/containerized environments with memory constraints.
-    ///     - 1-1024: Manual override - use specified number of concurrent operations
-    ///
-    /// Backward compatibility: the legacy positional signature
-    ///   processBatch(inputs, outputDir, format, quality?, fastMode?, concurrency?)
-    /// is still accepted for now but will be removed in a future major release.
-    #[napi(js_name = "processBatch", ts_return_type = "Promise<BatchResult[]>")]
-    pub fn process_batch(
-        &self,
-        env: Env,
-        inputs: Vec<String>,
-        output_dir: String,
-        options_or_format: Either<BatchOptions, String>,
-        quality: Option<f64>,
-        fast_mode: Option<bool>,
-        concurrency: Option<f64>,
-    ) -> Result<AsyncTask<BatchTask>> {
-        if output_dir.trim().is_empty() {
-            return Err(napi_err(
-                &env,
-                LazyImageError::invalid_argument(
-                    "outputDir",
-                    "<empty>",
-                    "output directory must not be empty",
-                ),
-            ));
-        }
-
-        let (format, quality, fast_mode, concurrency) = match options_or_format {
-            Either::A(options) => (
-                options.format,
-                options.quality,
-                options.fast_mode,
-                options.concurrency,
-            ),
-            Either::B(format) => (format, quality, fast_mode, concurrency),
-        };
-
-        let fast_mode = fast_mode.unwrap_or(false);
-        let quality = validation::sanitize_quality(quality).map_err(|e| napi_err(&env, e))?;
-        let concurrency =
-            validation::sanitize_concurrency(concurrency).map_err(|e| napi_err(&env, e))?;
-
-        let output_format = match OutputFormat::from_str_with_options(&format, quality, fast_mode) {
-            Ok(format) => format,
-            Err(_e) => {
-                let lazy_err = LazyImageError::unsupported_format(format.clone());
-                return Err(crate::error::napi_error_with_code(&env, lazy_err)?);
-            }
-        };
-        let ops = self.ops.clone();
-        let mut policy = self.metadata_policy.clone();
-        policy.apply_firewall(self.firewall.reject_metadata);
-        Ok(AsyncTask::new(BatchTask {
-            inputs,
-            output_dir,
-            ops,
-            format: output_format,
-            concurrency,
-            metadata_policy: policy,
-            auto_orient: self.auto_orient,
-            firewall: self.firewall.clone(),
-            #[cfg(feature = "napi")]
-            last_error: None,
-        }))
-    }
-
-    #[napi(
-        js_name = "processBatchWithMetrics",
-        ts_return_type = "Promise<BatchOutputWithMetrics>"
-    )]
-    pub fn process_batch_with_metrics(
-        &self,
-        env: Env,
-        inputs: Vec<String>,
-        output_dir: String,
-        options_or_format: Either<BatchOptions, String>,
-        quality: Option<f64>,
-        fast_mode: Option<bool>,
-        concurrency: Option<f64>,
-    ) -> Result<AsyncTask<BatchWithMetricsTask>> {
-        if output_dir.trim().is_empty() {
-            return Err(napi_err(
-                &env,
-                LazyImageError::invalid_argument(
-                    "outputDir",
-                    "<empty>",
-                    "output directory must not be empty",
-                ),
-            ));
-        }
-
-        let (format, quality, fast_mode, concurrency) = match options_or_format {
-            Either::A(options) => (
-                options.format,
-                options.quality,
-                options.fast_mode,
-                options.concurrency,
-            ),
-            Either::B(format) => (format, quality, fast_mode, concurrency),
-        };
-
-        let fast_mode = fast_mode.unwrap_or(false);
-        let quality = validation::sanitize_quality(quality).map_err(|e| napi_err(&env, e))?;
-        let concurrency =
-            validation::sanitize_concurrency(concurrency).map_err(|e| napi_err(&env, e))?;
-
-        let output_format = match OutputFormat::from_str_with_options(&format, quality, fast_mode) {
-            Ok(format) => format,
-            Err(_e) => {
-                let lazy_err = LazyImageError::unsupported_format(format.clone());
-                return Err(crate::error::napi_error_with_code(&env, lazy_err)?);
-            }
-        };
-        let ops = self.ops.clone();
-        let mut policy = self.metadata_policy.clone();
-        policy.apply_firewall(self.firewall.reject_metadata);
-
-        Ok(AsyncTask::new(BatchWithMetricsTask {
-            inputs,
-            output_dir,
-            ops,
-            format: output_format,
-            concurrency,
-            metadata_policy: policy,
-            auto_orient: self.auto_orient,
-            firewall: self.firewall.clone(),
-            #[cfg(feature = "napi")]
-            last_error: None,
-        }))
-    }
-}
-
-#[cfg(feature = "napi")]
-#[napi(object)]
-pub struct SanitizeOptions {
-    pub policy: Option<String>,
-}
-
-#[cfg(feature = "napi")]
-#[napi(object)]
-pub struct FirewallLimitOptions {
-    pub max_pixels: Option<f64>,
-    pub max_bytes: Option<f64>,
-    pub timeout_ms: Option<f64>,
-}
-
-#[cfg(feature = "napi")]
-#[napi(object)]
-pub struct Dimensions {
-    pub width: u32,
-    pub height: u32,
-}
-
-#[cfg(feature = "napi")]
-/// Result of applying a preset, contains recommended output settings
-#[napi(object)]
-pub struct PresetResult {
-    /// Recommended output format
-    pub format: String,
-    /// Recommended quality (None for PNG)
-    pub quality: Option<u8>,
-    /// Target width (None if aspect ratio preserved)
-    pub width: Option<u32>,
-    /// Target height (None if aspect ratio preserved)
-    pub height: Option<u32>,
-}
-
-#[cfg(feature = "napi")]
-#[napi(object)]
-pub struct BatchOptions {
-    /// Output format ("jpeg", "png", "webp", "avif")
-    pub format: String,
-    /// Optional quality (1-100), uses format default when omitted. Omit for PNG.
-    pub quality: Option<f64>,
-    /// Optional fast mode flag (JPEG only, default: false)
-    pub fast_mode: Option<bool>,
-    /// Optional number of parallel workers:
-    /// 0/undefined = auto-detect, 1-1024 = manual override
-    pub concurrency: Option<f64>,
-}
-
-// =============================================================================
-// INTERNAL IMPLEMENTATION
-// =============================================================================
-
-impl ImageEngine {
-    /// Lazily extract and return the ICC profile from the source data.
-    /// The extraction runs at most once; subsequent calls return the cached result.
-    fn icc_profile(&self) -> Option<&Arc<Vec<u8>>> {
-        self.icc_profile
-            .get_or_init(|| {
-                self.source
-                    .as_ref()
-                    .and_then(|s| s.as_bytes())
-                    .and_then(|bytes| extract_icc_profile_lossy(bytes).map(Arc::new))
-            })
-            .as_ref()
-    }
-
-    /// Lazily extract and return the raw EXIF data from the source data.
-    /// The extraction runs at most once; subsequent calls return the cached result.
-    fn exif_data(&self) -> Option<&Arc<Vec<u8>>> {
-        self.exif_data
-            .get_or_init(|| {
-                self.source
-                    .as_ref()
-                    .and_then(|s| s.as_bytes())
-                    .and_then(|bytes| extract_exif_raw(bytes).map(Arc::new))
-            })
-            .as_ref()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::engine::resize::fast_resize_owned;
-    #[allow(unused_imports)]
-    use image::GenericImageView;
-    use image::{DynamicImage, RgbImage};
-
-    fn create_test_image(width: u32, height: u32) -> DynamicImage {
-        DynamicImage::ImageRgb8(RgbImage::from_fn(width, height, |x, y| {
-            image::Rgb([(x % 256) as u8, (y % 256) as u8, 128])
-        }))
-    }
-
-    #[test]
-    fn fast_resize_owned_returns_error_instead_of_dummy_image() {
-        let img = create_test_image(1, 1);
-        let err = fast_resize_owned(img, 0, 10).expect_err("expected resize failure");
-        assert_eq!(err.source_dims, (1, 1));
-        assert_eq!(err.target_dims, (0, 10));
-        assert!(err.reason.contains("invalid dimensions"));
-    }
-}
